// src/app/models/receivers.model.ts
export interface Receivers {
    firstName?: string;
    middleName?: string;
    lastName?: string;
    country?: string;
    countryCode?: number;
    type?: string;
    phoneNumber?: number;
    bank?: string;
    accNo?: number;
    txn?: string;
    sendAmount?: number;
    receivedAmount?: number;
    payIn?: string;
    payOut?: string;
    payMethod?: string;
}